SceAppMgr: implement sceAppMgrSaveDataDataSave and sceAppMgrSaveDataDataRemove
